+++ src/mailman/rest/addresses.py
@@ -211,11 +211,20 @@ class UserAddresses(_AddressBase):
         except InvalidEmailAddressError:
             bad_request(response, b'Invalid email address')
         except ExistingAddressError:
-            bad_request(response, b'Address already exists')
+            # Check if the address is not linked to any user, link it to the
+            # current user and return it.  Since we're linking to an existing
+            # address, ignore any given display_name attribute.
+            address = user_manager.get_address(validator(request)['email'])
+            if address.user is None:
+                address.user = self._user
+                location = self.path_to('addresses/{}'.format(address.email))
+                created(response, location)
+            else:
+                bad_request(response, 'Address belongs to other user.')
         else:
             # Link the address to the current user and return it.
             address.user = self._user
-            location = self.path_to('addresses/{0}'.format(address.email))
+            location = self.path_to('addresses/{}'.format(address.email))
             created(response, location)